proper content-disposition for index files
Typically, http.ServeContent can infer the MIME type from the first few bytes of the file. However, if the html file contains a byte order mark, then ServeContent falls back on the filename. If there is no filename, the MIME type defaults to text/plain. Thus, for index files, we should manually specify the file name so that the MIME type is correct when there is a byte-order mark.